Redirect debugging (#27)
