+++ frontend/src/pages/TeamsLayout.tsx
@@ -1,8 +1,10 @@
-import { Outlet } from "react-router-dom";
+import { Outlet, useLocation } from "react-router-dom";
 
 export default function TeamsLayout() {
+  const location = useLocation();
+  
   return (
-    <main>
+    <main key={location.pathname}>
       <Outlet />
     </main>
   );
